stream-processor/requirements.txt
kafka-python
psycopg2-binary
sqlalchemy
